TS-35786 move arguments to own class
