+++ src/test/java/com/teamscale/upload/IntegrationTestBase.java
@@ -2,15 +2,11 @@ package com.teamscale.upload;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
 
@@ -29,16 +25,15 @@ import com.teamscale.upload.utils.SecretUtils;
 public abstract class IntegrationTestBase {
 
 	private static final int MOCK_TEAMSCALE_PORT = 24398;
-	private static final String TEAMSCALE_TEST_USER = "teamscale-upload-build-test-user";
 
 
 	/**
 	 * Executes the generated teamscale-upload distribution with the given arguments.
 	 */
-	protected abstract ProcessUtils.ProcessResult runUploader(Arguments arguments);
+	protected abstract ProcessUtils.ProcessResult runUploader(TeamscaleUploadArguments arguments);
 	@Test
 	public void wrongAccessKey() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withAccessKey("wrong-accesskey_"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withAccessKey("wrong-accesskey_"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("You provided incorrect credentials");
@@ -48,7 +43,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void incorrectUrl() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withUrl("no-protocol:9999"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withUrl("no-protocol:9999"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			// the command line library we use adjusts the word spacing based on the
@@ -61,7 +56,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void unresolvableUrl() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withUrl("http://domain.invalid:9999"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withUrl("http://domain.invalid:9999"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("could not be resolved");
@@ -71,7 +66,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void unreachableUrl() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withUrl("http://localhost:9999"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withUrl("http://localhost:9999"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("refused a connection");
@@ -81,7 +76,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void wrongUser() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withUser("wrong-user_"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withUser("wrong-user_"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("You provided incorrect credentials");
@@ -91,7 +86,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void wrongProject() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withProject("wrong-project_"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withProject("wrong-project_"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("The project")
@@ -103,7 +98,7 @@ public abstract class IntegrationTestBase {
 	@Test
 	public void timeoutTooSmall() {
 		ProcessUtils.ProcessResult result = runUploader(
-				new Arguments().withUrl("http://localhost:9999").withTimeoutInSeconds("0"));
+				new TeamscaleUploadArguments().withUrl("http://localhost:9999").withTimeoutInSeconds("0"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			// we ignore whitespace because the string is formatted differently coming out of the jlink dist than from graalvm (different line width)
@@ -114,7 +109,7 @@ public abstract class IntegrationTestBase {
 	@Test
 	public void timeoutIsNotANumber() {
 		ProcessUtils.ProcessResult result = runUploader(
-				new Arguments().withUrl("http://localhost:9999").withTimeoutInSeconds("foo"));
+				new TeamscaleUploadArguments().withUrl("http://localhost:9999").withTimeoutInSeconds("foo"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			// we ignore whitespace because the string is formatted differently coming out of the jlink dist than from graalvm (different line width)
@@ -129,7 +124,7 @@ public abstract class IntegrationTestBase {
 	 */
 	@Test
 	public void unknownRevision() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withCommit("doesnt-exist"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withCommit("doesnt-exist"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).doesNotContain("The project")
@@ -142,7 +137,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void patternMatchesNothing() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withPattern("**/matches.nothing"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withPattern("**/matches.nothing"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("The pattern")
@@ -153,7 +148,7 @@ public abstract class IntegrationTestBase {
 	@Test
 	public void insufficientPermissions() {
 		ProcessUtils.ProcessResult result = runUploader(
-				new Arguments().withUser("teamscale-upload-build-test-user-no-permissions").withAccessKey("ruG8MKMbLunyLooB7SlfkEATCISSWDKy"));
+				new TeamscaleUploadArguments().withUser("teamscale-upload-build-test-user-no-permissions").withAccessKey("ruG8MKMbLunyLooB7SlfkEATCISSWDKy"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("is not allowed to upload data to the Teamscale project");
@@ -163,7 +158,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void successfulSingleFormatUpload() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments());
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments());
 		assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 		assertThatOSCertificatesWereImported(result);
 	}
@@ -171,7 +166,7 @@ public abstract class IntegrationTestBase {
 	@Test
 	public void successfulMultiFormatUpload() {
 		ProcessUtils.ProcessResult result = runUploader(
-				new Arguments().withInput("src/test/resources/coverage_files/input_file"));
+				new TeamscaleUploadArguments().withInput("src/test/resources/coverage_files/input_file"));
 		assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 		assertThatOSCertificatesWereImported(result);
 	}
@@ -179,7 +174,7 @@ public abstract class IntegrationTestBase {
 	@Test
 	public void testDefaultMessage() {
 		try (TeamscaleMockServer server = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT)) {
-			ProcessUtils.ProcessResult result = runUploader(new Arguments()
+			ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments()
 					.withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withAdditionalMessageLine("Build ID: 1234"));
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 			assertThat(server.sessions).hasSize(1).first().extracting(this::extractNormalizedMessage)
@@ -194,7 +189,7 @@ public abstract class IntegrationTestBase {
 	public void testTimeoutSmallerThanRequestTime() {
 		try (TeamscaleMockServer ignored = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT, false, 2)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withTimeoutInSeconds("1"));
+					new TeamscaleUploadArguments().withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withTimeoutInSeconds("1"));
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput())
 					.isNotZero();
 			assertThat(result.errorOutput).contains("Request timeout reached.");
@@ -206,7 +201,7 @@ public abstract class IntegrationTestBase {
 	public void testTimeoutGreaterThanRequestTime() {
 		try (TeamscaleMockServer ignored = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT, false, 2)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withTimeoutInSeconds("3"));
+					new TeamscaleUploadArguments().withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withTimeoutInSeconds("3"));
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 			assertThatOSCertificatesWereImported(result);
 		}
@@ -214,7 +209,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void mustRejectTimestampPassedInSeconds() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withTimestamp("master:1606764633"));
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withTimestamp("master:1606764633"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput())
 					.isNotZero();
@@ -227,7 +222,7 @@ public abstract class IntegrationTestBase {
 	public void selfSignedCertificateShouldBeAcceptedWithInsecureFlag() {
 		try (TeamscaleMockServer server = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT, true)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT).withInsecure());
+					new TeamscaleUploadArguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT).withInsecure());
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 			assertThat(server.sessions).hasSize(1);
 			assertThatOSCertificatesWereImported(result);
@@ -238,7 +233,7 @@ public abstract class IntegrationTestBase {
 	public void selfSignedCertificateShouldNotBeAcceptedByDefault() {
 		try (TeamscaleMockServer ignored = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT, true)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT));
+					new TeamscaleUploadArguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT));
 			assertSoftlyThat(softly -> {
 				softly.assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput())
 						.isNotZero();
@@ -252,7 +247,7 @@ public abstract class IntegrationTestBase {
 	public void printStackTraceForKnownErrorsOnlyWhenRequested() {
 		try (TeamscaleMockServer ignored = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT, true)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT));
+					new TeamscaleUploadArguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT));
 			assertSoftlyThat(softly -> {
 				softly.assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput())
 						.isNotZero();
@@ -263,7 +258,7 @@ public abstract class IntegrationTestBase {
 
 		try (TeamscaleMockServer ignored = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT, true)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT).withStackTrace());
+					new TeamscaleUploadArguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT).withStackTrace());
 			assertSoftlyThat(softly -> {
 				softly.assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput())
 						.isNotZero();
@@ -278,7 +273,7 @@ public abstract class IntegrationTestBase {
 	public void selfSignedCertificateShouldBeAcceptedWhenKeystoreIsUsed() {
 		try (TeamscaleMockServer server = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT, true)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT).withKeystore().withStackTrace());
+					new TeamscaleUploadArguments().withUrl("https://localhost:" + MOCK_TEAMSCALE_PORT).withKeystore().withStackTrace());
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 			assertThat(server.sessions).hasSize(1);
 			assertThatOSCertificatesWereImported(result);
@@ -289,7 +284,7 @@ public abstract class IntegrationTestBase {
 	public void mustGuessRevision() {
 		try (TeamscaleMockServer server = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT)) {
 			ProcessUtils.ProcessResult result = runUploader(
-					new Arguments().withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withAutoDetectCommit());
+					new TeamscaleUploadArguments().withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withAutoDetectCommit());
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 			assertThat(server.sessions).hasSize(1);
 			assertThat(server.sessions.get(0).revisionOrTimestamp).hasSize(40); // size of a git SHA1
@@ -309,7 +304,7 @@ public abstract class IntegrationTestBase {
 			tempFilePath = Paths.get(temporaryFileName);
 			Files.writeString(tempFilePath, System.getenv(SecretUtils.TEAMSCALE_ACCESS_KEY_ENVIRONMENT_VARIABLE));
 
-			ProcessUtils.ProcessResult result = runUploader(new Arguments().withAccessKeyViaStdin(temporaryFileName));
+			ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withAccessKeyViaStdin(temporaryFileName));
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 			assertThatOSCertificatesWereImported(result);
 		} finally {
@@ -323,7 +318,7 @@ public abstract class IntegrationTestBase {
 	@Test
 	public void testIncorrectAccessKeyFromStdIn() {
 		ProcessUtils.ProcessResult result = runUploader(
-				new Arguments().withAccessKeyViaStdin("src/test/resources/incorrect_access_key.txt"));
+				new TeamscaleUploadArguments().withAccessKeyViaStdin("src/test/resources/incorrect_access_key.txt"));
 		assertSoftlyThat(softly -> {
 			softly.assertThat(result.exitCode).isNotZero();
 			softly.assertThat(result.errorOutput).contains("You provided incorrect credentials");
@@ -337,7 +332,7 @@ public abstract class IntegrationTestBase {
 	 */
 	@Test
 	public void testCorrectAccessWithKeyFromEnvironmentVariable() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withoutAccessKeyInOption());
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withoutAccessKeyInOption());
 		assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
 		assertThatOSCertificatesWereImported(result);
 	}
@@ -346,7 +341,7 @@ public abstract class IntegrationTestBase {
 	@EnabledOnOs(OS.MAC)
 	public void testXCResultConversion() throws IOException {
 		try (TeamscaleMockServer server = new TeamscaleMockServer(MOCK_TEAMSCALE_PORT)) {
-			ProcessUtils.ProcessResult result = runUploader(new Arguments()
+			ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments()
 					.withPattern("src/test/resources/coverage_files/output.xcresult.tar.gz").withFormat("XCODE")
 					.withUrl("http://localhost:" + MOCK_TEAMSCALE_PORT).withAutoDetectCommit());
 			assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.getOutputAndErrorOutput()).isZero();
@@ -371,7 +366,7 @@ public abstract class IntegrationTestBase {
 	@Test
 	public void successfulUploadWithRepository() {
 		ProcessUtils.ProcessResult result = runUploader(
-				new Arguments().withRepository("cqse/teamscale-upload").withPartition("NativeImageIT > TestRepository")
+				new TeamscaleUploadArguments().withRepository("cqse/teamscale-upload").withPartition("NativeImageIT > TestRepository")
 						.withCommit("3758a3a6c2d62ab787574f869b2352480c6f0c10"));
 		assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.errorOutput).isZero();
 		assertThatOSCertificatesWereImported(result);
@@ -379,7 +374,7 @@ public abstract class IntegrationTestBase {
 
 	@Test
 	public void successfulUploadWithMoveToLastCommit() {
-		ProcessUtils.ProcessResult result = runUploader(new Arguments().withMoveToLastCommit());
+		ProcessUtils.ProcessResult result = runUploader(new TeamscaleUploadArguments().withMoveToLastCommit());
 		assertThat(result.exitCode).describedAs("Stderr and stdout: " + result.errorOutput).isZero();
 		assertThatOSCertificatesWereImported(result);
 	}
@@ -411,202 +406,4 @@ public abstract class IntegrationTestBase {
 		return session.message.replaceAll("uploaded from .*", "uploaded from HOSTNAME").replaceAll("uploaded at .*",
 				"uploaded at DATE");
 	}
-
-	private static String getAccessKeyFromCi() {
-		String accessKey = System.getenv(SecretUtils.TEAMSCALE_ACCESS_KEY_ENVIRONMENT_VARIABLE);
-		if (accessKey == null) {
-			return "not-a-ci-build";
-		}
-		return accessKey;
-	}
-
-	/**
-	 * Arguments for an execution of the teamscale-upload executable.
-	 */
-	static class Arguments {
-		private String partition = "NativeImageIT";
-		private String url = "https://cqse.teamscale.io/";
-		private String user = TEAMSCALE_TEST_USER;
-		private String accessKey = getAccessKeyFromCi();
-		private String project = "teamscale-upload";
-		private String format = "simple";
-		private String pattern = "src/test/resources/coverage_files\\*.simple";
-		private String input = null;
-		private boolean insecure = false;
-		private boolean useKeystore = false;
-		private boolean autoDetectCommit = false;
-		private String timestamp = "master:HEAD";
-		private String commit = null;
-		private String repository = null;
-		private String additionalMessageLine = null;
-		private boolean stackTrace = false;
-
-		/**
-		 * The file from which the teamscale-upload executable should draw its stdin.
-		 */
-		File stdinFile = null;
-		private boolean moveToLastCommit = false;
-		private String timeoutInSeconds = null;
-
-		private Arguments withFormat(String format) {
-			this.format = format;
-			return this;
-		}
-
-		private Arguments withPattern(String pattern) {
-			this.pattern = pattern;
-			return this;
-		}
-
-		private Arguments withInsecure() {
-			this.insecure = true;
-			return this;
-		}
-
-		private Arguments withCommit(String commit) {
-			this.commit = commit;
-			return this;
-		}
-
-		private Arguments withTimestamp(String timestamp) {
-			this.timestamp = timestamp;
-			return this;
-		}
-
-		private Arguments withRepository(String repository) {
-			this.repository = repository;
-			return this;
-		}
-
-		private Arguments withAutoDetectCommit() {
-			this.autoDetectCommit = true;
-			return this;
-		}
-
-		private Arguments withKeystore() {
-			this.useKeystore = true;
-			return this;
-		}
-
-		private Arguments withAdditionalMessageLine(String line) {
-			this.additionalMessageLine = line;
-			return this;
-		}
-
-		private Arguments withUrl(String url) {
-			this.url = url;
-			return this;
-		}
-
-		private Arguments withStackTrace() {
-			this.stackTrace = true;
-			return this;
-		}
-
-		private Arguments withAccessKey(String accessKey) {
-			this.accessKey = accessKey;
-			return this;
-		}
-
-		/**
-		 * No access key is specified as option. The key which is specified in the
-		 * environment variable should be used instead.
-		 */
-		private Arguments withoutAccessKeyInOption() {
-			this.accessKey = null;
-			return this;
-		}
-
-		private Arguments withAccessKeyViaStdin(String stdinFilePath) {
-			this.accessKey = "-";
-			// If the access key is set to '-', we need to pipe the key from a file via
-			// stdin.
-			this.stdinFile = new File(stdinFilePath);
-			return this;
-		}
-
-		private Arguments withUser(String user) {
-			this.user = user;
-			return this;
-		}
-
-		private Arguments withProject(String project) {
-			this.project = project;
-			return this;
-		}
-
-		private Arguments withInput(String input) {
-			this.input = input;
-			return this;
-		}
-
-		private Arguments withPartition(String partition) {
-			this.partition = partition;
-			return this;
-		}
-
-		private Arguments withMoveToLastCommit() {
-			this.moveToLastCommit = true;
-			return this;
-		}
-
-		private Arguments withTimeoutInSeconds(String timeoutInSeconds) {
-			this.timeoutInSeconds = timeoutInSeconds;
-			return this;
-		}
-
-		/**
-		 * Assembles the command that invokes the given teamscale-upload executable.
-		 */
-		String[] toCommand(String executable) {
-			List<String> command = new ArrayList<>(
-					Arrays.asList(executable, "-s", url, "-u", user, "-f", format, "-p", project, "-t", partition));
-			if (accessKey != null) {
-				command.add("-a");
-				command.add(accessKey);
-			}
-			if (input != null) {
-				command.add("-i");
-				command.add(input);
-			}
-			command.add(pattern);
-			if (insecure) {
-				command.add("--insecure");
-			}
-			if (useKeystore) {
-				command.add("--trusted-keystore");
-				command.add(TeamscaleMockServer.TRUSTSTORE.getAbsolutePath() + ";password");
-			}
-			if (additionalMessageLine != null) {
-				command.add("--append-to-message");
-				command.add(additionalMessageLine);
-			}
-			if (stackTrace) {
-				command.add("--stacktrace");
-			}
-			if (moveToLastCommit) {
-				command.add("--move-to-last-commit");
-			}
-
-			if (commit != null) {
-				command.add("--commit");
-				command.add(commit);
-			} else if (!autoDetectCommit) {
-				command.add("--branch-and-timestamp");
-				command.add(timestamp);
-			}
-
-			if (repository != null) {
-				command.add("--repository");
-				command.add(repository);
-			}
-			if (timeoutInSeconds != null) {
-				command.add("--timeout");
-				command.add(timeoutInSeconds);
-			}
-
-			return command.toArray(new String[0]);
-		}
-	}
-
 }
